Version 1.04

+++ cron/receive_bitcoin.php
@@ -2,7 +2,7 @@
 <?php
 echo "Beginning Receive Bitcoin processing...".PHP_EOL;
 
-include 'cfg.php';
+include 'common.php';
 
 $CFG->session_active = true;
 $transactions_dir = $CFG->dirroot.'transactions/';
@@ -75,11 +75,11 @@ foreach ($transactions as $t_id) {
 		continue;
 	
 	$transaction = $bitcoin->gettransaction($t_id);
-	if ($transaction_log[$t_id]) {
+	if (!empty($transaction_log[$t_id])) {
 		unlink($transactions_dir.$t_id);
 		continue;
 	}
-	if (!$transaction['details'])
+	if (empty($transaction['details']))
 		continue;
 	
 	//$raw = $bitcoin->decoderawtransaction($bitcoin->getrawtransaction($t_id));
@@ -91,16 +91,16 @@ foreach ($transactions as $t_id) {
 	
 	foreach ($transaction['details'] as $detail) {
 		if ($detail['category'] == 'receive') {
-			$user_id = $addresses[$detail['address']];
-			$request_id = $requests[$transaction['txid']];
+			$user_id = (!empty($addresses[$detail['address']])) ? $addresses[$detail['address']] : false;
+			$request_id = (!empty($requests[$transaction['txid']])) ? $requests[$transaction['txid']] : false;
 			
-			if ($system[$detail['address']] == 'Y') {
+			if (!empty($system[$detail['address']]) && $system[$detail['address']] == 'Y') {
 				if ($transaction['confirmations'] > 0) {
 					$hot_wallet_in = $detail['amount'];
 				}
 				continue;
 			}
-			elseif ($system[$detail['address']] == 'N') {
+			elseif (!empty($system[$detail['address']]) && $system[$detail['address']] == 'N') {
 				unlink($transactions_dir.$t_id);
 				break;
 			}
@@ -189,6 +189,20 @@ $reserve_surplus = $hot_wallet - $reserve_balance - $pending_withdrawals - $CFG-
 
 echo 'Reserve surplus: '.sprintf("%.8f", $reserve_surplus).PHP_EOL;
 
+/*
+if ($total_received > 0 || $reserve_surplus > $CFG->bitcoin_reserve_min) {
+	$ch = curl_init($CFG->hv_addr);
+	curl_setopt($ch,CURLOPT_RETURNTRANSFER,true);
+	curl_setopt($ch,CURLOPT_POSTFIELDS,array('key'=>$CFG->hv_key));
+	curl_setopt($ch,CURLOPT_FRESH_CONNECT,TRUE);
+	$result1 = curl_exec($ch);
+	$result = json_decode($result1,true);
+	curl_close($ch);
+	$havelock_warm_wallet = $result['address'];
+}
+*/
+$havelock_warm_wallet = $CFG->bitcoin_warm_wallet_address;
+
 if ($total_received > 0) {
 	if (!$status) {
 		echo 'Error: Could not get status.'.PHP_EOL;
@@ -200,10 +214,10 @@ if ($total_received > 0) {
 	$updated = db_update('status',1,array('hot_wallet_btc'=>$hot_wallet,'total_btc'=>$total_btc));
 	
 	//$warm_wallet_a = BitcoinAddresses::getWarmWallet();
-	$warm_wallet_a['address'] = $CFG->bitcoin_warm_wallet_address;
+	$warm_wallet_a['address'] = $havelock_warm_wallet;
 	$hot_wallet_a = BitcoinAddresses::getHotWallet();
 	
-	if ($reserve_surplus > $CFG->bitcoin_reserve_min) {
+	if ($reserve_surplus > $CFG->bitcoin_reserve_min && $havelock_warm_wallet) {
 		$bitcoin->walletpassphrase($CFG->bitcoin_passphrase,3);
 		$response = $bitcoin->sendfrom($CFG->bitcoin_accountname,$warm_wallet_a['address'],floatval($reserve_surplus));
 		$transferred = 0;
@@ -227,9 +241,9 @@ if ($total_received > 0) {
 		}
 	}
 }
-elseif ($reserve_surplus > $CFG->bitcoin_reserve_min) {
+elseif ($reserve_surplus > $CFG->bitcoin_reserve_min && $havelock_warm_wallet) {
 	//$warm_wallet_a = BitcoinAddresses::getWarmWallet();
-	$warm_wallet_a['address'] = $CFG->bitcoin_warm_wallet_address;
+	$warm_wallet_a['address'] = $havelock_warm_wallet;
 	$hot_wallet_a = BitcoinAddresses::getHotWallet();
 	
 	$bitcoin->settxfee(0.00);
